backend/requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.35.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
Pillow==10.1.0
scikit-learn==1.3.2
tqdm==4.66.1
matplotlib==3.8.2
